dom/base/test/test_blob_fragment_and_query.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Blob URI with fragments</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script>

var  blob = new Blob(['hello world']);
ok(blob, "We have a blob.");

var url = URL.createObjectURL(blob);
ok(url, "We have a URI");

var tests = [
  url,
  url + "?aa",
  url + "#bb",
  url + "?cc#dd",
  url + "#ee?ff",
];

function runTest() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();

  var xhr = new XMLHttpRequest();
  xhr.open('GET', test);
  xhr.onload = function() {
    is(xhr.responseText, 'hello world', 'URL: ' + test);
    runTest();
  }
  xhr.send();
}

SimpleTest.waitForExplicitFinish();
runTest();

  </script>
</body>
</html>
